my_agent/requirements.txt
langgraph
langchain_anthropic
tavily-python
langchain_community
langchain_openai
langgraph-cli[inmem]
langchain-mcp-adapters
mcp
python-dotenv
langchain
